Fixed Profile
I fixed the profile thingy and organized the components stuff into a components folder
